fix(csc): harden cache and connection lifecycle
Serialize cache publication with connection lifecycle changes and evict entries when tracking coverage is lost.
