--- csc_commands_test.go
@@ -1,393 +1,0 @@
-package redis
-
-import (
-	"bytes"
-	"context"
-	"testing"
-)
-
-// helper to create a Cmd with the given args.
-func makeCmd(args ...interface{}) Cmder {
-	return NewCmd(context.Background(), args...)
-}
-
-// --- isCacheable -----------------------------------------------------------
-
-func TestIsCacheable_AllowedCommands(t *testing.T) {
-	allowed := []string{
-		"GET", "MGET", "HGET", "HMGET", "HGETALL",
-		"HKEYS", "HVALS", "HLEN", "HEXISTS", "HSTRLEN",
-		"LINDEX", "LLEN", "LPOS", "LRANGE",
-		"SCARD", "SISMEMBER", "SMEMBERS", "SMISMEMBER",
-		"SDIFF", "SINTER", "SINTERCARD", "SUNION",
-		"ZCARD", "ZCOUNT", "ZLEXCOUNT", "ZMSCORE",
-		"ZRANGE", "ZRANGEBYLEX", "ZRANGEBYSCORE",
-		"ZRANK", "ZREVRANGE", "ZREVRANGEBYLEX",
-		"ZREVRANGEBYSCORE", "ZREVRANK", "ZSCORE",
-		"ZDIFF", "ZINTER", "ZUNION",
-		"STRLEN", "GETBIT", "GETRANGE", "SUBSTR",
-		"BITCOUNT", "BITFIELD_RO", "BITPOS",
-		"EXISTS", "TYPE", "SORT_RO", "LCS",
-		"GEODIST", "GEOHASH", "GEOPOS", "GEOSEARCH",
-		"GEORADIUSBYMEMBER_RO", "GEORADIUS_RO",
-		"XLEN", "XRANGE", "XREVRANGE",
-		"JSON.GET", "JSON.MGET", "JSON.ARRINDEX", "JSON.ARRLEN",
-		"JSON.OBJKEYS", "JSON.OBJLEN", "JSON.RESP",
-		"JSON.STRLEN", "JSON.TYPE",
-		"TS.GET", "TS.INFO", "TS.RANGE", "TS.REVRANGE",
-	}
-	for _, name := range allowed {
-		// Use lower-case name as first arg (matching how go-redis sends commands)
-		cmd := makeCmd(name, "mykey")
-		if !isCacheable(cmd) {
-			t.Errorf("expected %q to be cacheable", name)
-		}
-	}
-}
-
-func TestIsCacheable_CaseInsensitive(t *testing.T) {
-	for _, name := range []string{"get", "Get", "GET", "gEt"} {
-		cmd := makeCmd(name, "k")
-		if !isCacheable(cmd) {
-			t.Errorf("expected %q to be cacheable (case-insensitive)", name)
-		}
-	}
-}
-
-func TestIsCacheable_WriteCommandsRejected(t *testing.T) {
-	writes := []string{"SET", "DEL", "HSET", "LPUSH", "SADD", "ZADD", "EXPIRE", "FLUSHDB"}
-	for _, name := range writes {
-		cmd := makeCmd(name, "k")
-		if isCacheable(cmd) {
-			t.Errorf("expected %q to NOT be cacheable", name)
-		}
-	}
-}
-
-func TestIsCacheable_XReadRejected(t *testing.T) {
-	// XREAD supports BLOCK and state-relative $/+ IDs, so it must not be cached.
-	cmd := makeCmd("XREAD", "COUNT", "5", "STREAMS", "s", "0")
-	if isCacheable(cmd) {
-		t.Error("expected XREAD to NOT be cacheable")
-	}
-}
-
-// TestExtractRedisKeys_WireFaithfulTypesOnly: the invalidation index must hold
-// keys exactly as proto.Writer sends them, or the server's invalidation pushes
-// never match and stale entries are served forever. Types whose fmt.Sprint
-// rendering diverges from the wire form (pointers, bools, durations, floats...)
-// must make extraction fail (want nil) so the command is served uncached.
-func TestExtractRedisKeys_WireFaithfulTypesOnly(t *testing.T) {
-	key := "real-key"
-	cases := []struct {
-		name string
-		cmd  Cmder
-		want []string
-	}{
-		{"string key", makeCmd("get", "k"), []string{"k"}},
-		{"[]byte key", makeCmd("get", []byte("k")), []string{"k"}},
-		{"int key", makeCmd("get", 123), []string{"123"}},
-		{"uint64 key", makeCmd("get", uint64(7)), []string{"7"}},
-		{"pointer key", makeCmd("get", &key), nil},
-		{"bool key", makeCmd("get", true), nil},
-		{"float key", makeCmd("get", 1.5), nil},
-		// Multi-key commands: one divergent key poisons the whole extraction.
-		{"mget with pointer key", makeCmd("mget", "a", &key, "b"), nil},
-		{"mget with string keys", makeCmd("mget", "a", "b"), []string{"a", "b"}},
-	}
-	for _, tc := range cases {
-		got := extractRedisKeys(tc.cmd)
-		if len(got) != len(tc.want) {
-			t.Errorf("%s: got %v, want %v", tc.name, got, tc.want)
-			continue
-		}
-		for i := range got {
-			if got[i] != tc.want[i] {
-				t.Errorf("%s: got %v, want %v", tc.name, got, tc.want)
-				break
-			}
-		}
-	}
-}
-
-func TestIsCacheable_XPendingRejected(t *testing.T) {
-	// XPENDING's extended form returns wall-clock-relative idle times and its
-	// IDLE filter is time-dependent, so it must not be cached.
-	for _, cmd := range []Cmder{
-		makeCmd("XPENDING", "s", "grp"),
-		makeCmd("XPENDING", "s", "grp", "IDLE", "9000", "-", "+", "10"),
-	} {
-		if isCacheable(cmd) {
-			t.Errorf("expected %v to NOT be cacheable", cmd.Args())
-		}
-	}
-}
-
-func TestIsCacheable_KeylessCommandRejected(t *testing.T) {
-	// PING has no keys; even if someone added it to the allow-list it
-	// should be rejected because cmdFirstKeyPos returns 0.
-	cmd := makeCmd("ping")
-	if isCacheable(cmd) {
-		t.Error("expected keyless command PING to NOT be cacheable")
-	}
-}
-
-func TestIsCacheable_RawWriteToRejected(t *testing.T) {
-	cmd := NewRawWriteToCmd(context.Background(), &bytes.Buffer{}, "get", "k")
-	if isCacheable(cmd) {
-		t.Fatal("RawWriteToCmd must bypass CSC to preserve direct streaming")
-	}
-}
-
-func TestIsSelectCmd(t *testing.T) {
-	for _, cmd := range []Cmder{
-		makeCmd("select", 1),
-		makeCmd("SELECT", 1),
-		makeCmd([]byte("select"), 1),
-	} {
-		if !isSelectCmd(cmd) {
-			t.Errorf("expected %v to match SELECT", cmd.Args())
-		}
-	}
-	for _, cmd := range []Cmder{
-		makeCmd("get", "select"),
-		makeCmd("swapdb", 0, 1),
-	} {
-		if isSelectCmd(cmd) {
-			t.Errorf("expected %v not to match SELECT", cmd.Args())
-		}
-	}
-}
-
-func TestCSCStateCommandMatchers(t *testing.T) {
-	for _, cmd := range []Cmder{
-		makeCmd("auth", "password"),
-		makeCmd("AUTH", "user", "password"),
-		makeCmd([]byte("auth"), "password"),
-	} {
-		if !isAuthCmd(cmd) {
-			t.Errorf("expected %v to match AUTH", cmd.Args())
-		}
-	}
-	if isAuthCmd(makeCmd("get", "auth")) {
-		t.Fatal("GET auth must not match AUTH")
-	}
-
-	for _, cmd := range []Cmder{
-		makeCmd("hello", 2),
-		makeCmd("HELLO", 3),
-		makeCmd([]byte("hello"), []byte("2")),
-	} {
-		if !isProtocolChangingHelloCmd(cmd) {
-			t.Errorf("expected %v to match state-changing HELLO", cmd.Args())
-		}
-	}
-	for _, cmd := range []Cmder{
-		makeCmd("hello"),
-		makeCmd("get", "hello"),
-	} {
-		if isProtocolChangingHelloCmd(cmd) {
-			t.Errorf("expected %v not to match state-changing HELLO", cmd.Args())
-		}
-	}
-
-	for _, cmd := range []Cmder{
-		makeCmd("reset"),
-		makeCmd("RESET"),
-		makeCmd([]byte("reset")),
-	} {
-		if !isResetCmd(cmd) {
-			t.Errorf("expected %v to match RESET", cmd.Args())
-		}
-	}
-	if isResetCmd(makeCmd("config", "resetstat")) {
-		t.Fatal("CONFIG RESETSTAT must not match RESET")
-	}
-}
-
-func TestIsCacheable_EmptyArgs(t *testing.T) {
-	cmd := makeCmd()
-	if isCacheable(cmd) {
-		t.Error("expected empty command to NOT be cacheable")
-	}
-}
-
-// --- buildCacheKey ---------------------------------------------------------
-
-func TestBuildCacheKey_SimpleGet(t *testing.T) {
-	cmd := makeCmd("GET", "foo")
-	key, ok := buildCacheKey(cmd)
-	if !ok || key == "" {
-		t.Fatal("expected non-empty cache key")
-	}
-	// Same command must produce identical keys.
-	if key2, _ := buildCacheKey(makeCmd("GET", "foo")); key != key2 {
-		t.Errorf("identical commands produced different keys: %q vs %q", key, key2)
-	}
-}
-
-func TestBuildCacheKey_DifferentArgsDiffer(t *testing.T) {
-	k1, _ := buildCacheKey(makeCmd("GET", "foo"))
-	k2, _ := buildCacheKey(makeCmd("GET", "bar"))
-	if k1 == k2 {
-		t.Error("different keys must produce different cache keys")
-	}
-}
-
-func TestBuildCacheKey_CollisionSafety(t *testing.T) {
-	// "a|b" as one arg vs "a" and "b" as two args must differ.
-	k1, _ := buildCacheKey(makeCmd("GET", "a|b"))
-	k2, _ := buildCacheKey(makeCmd("GET", "a", "b"))
-	if k1 == k2 {
-		t.Error("length-prefixing should prevent separator collision")
-	}
-}
-
-func TestBuildCacheKey_BinaryData(t *testing.T) {
-	cmd := makeCmd("GET", []byte{0x00, 0x01, 0xff})
-	key, ok := buildCacheKey(cmd)
-	if !ok || key == "" {
-		t.Fatal("expected non-empty cache key for binary argument")
-	}
-}
-
-func TestBuildCacheKey_MultiKey(t *testing.T) {
-	k1, _ := buildCacheKey(makeCmd("MGET", "a", "b"))
-	k2, _ := buildCacheKey(makeCmd("MGET", "a", "b", "c"))
-	if k1 == k2 {
-		t.Error("different arg counts must produce different cache keys")
-	}
-}
-
-func TestBuildCacheKey_EmptyArgs(t *testing.T) {
-	cmd := makeCmd()
-	if key, ok := buildCacheKey(cmd); ok || key != "" {
-		t.Errorf("expected empty cache key for no-args command, got %q (ok=%v)", key, ok)
-	}
-}
-
-// --- extractRedisKeys ------------------------------------------------------
-
-func TestExtractRedisKeys_SingleKey(t *testing.T) {
-	cmd := makeCmd("GET", "mykey")
-	keys := extractRedisKeys(cmd)
-	if len(keys) != 1 || keys[0] != "mykey" {
-		t.Errorf("expected [mykey], got %v", keys)
-	}
-}
-
-func TestExtractRedisKeys_SingleKeyWithExtraArgs(t *testing.T) {
-	// LRANGE has one key followed by start/stop — only the key should be extracted.
-	cmd := makeCmd("LRANGE", "mylist", "0", "10")
-	keys := extractRedisKeys(cmd)
-	if len(keys) != 1 || keys[0] != "mylist" {
-		t.Errorf("LRANGE: expected [mylist], got %v", keys)
-	}
-
-	// HGET has one key followed by a field name.
-	cmd = makeCmd("HGET", "myhash", "field1")
-	keys = extractRedisKeys(cmd)
-	if len(keys) != 1 || keys[0] != "myhash" {
-		t.Errorf("HGET: expected [myhash], got %v", keys)
-	}
-
-	// ZCOUNT has one key followed by min/max.
-	cmd = makeCmd("ZCOUNT", "myset", "-inf", "+inf")
-	keys = extractRedisKeys(cmd)
-	if len(keys) != 1 || keys[0] != "myset" {
-		t.Errorf("ZCOUNT: expected [myset], got %v", keys)
-	}
-
-	// GETRANGE has one key followed by start/end offsets.
-	cmd = makeCmd("GETRANGE", "mystr", "0", "5")
-	keys = extractRedisKeys(cmd)
-	if len(keys) != 1 || keys[0] != "mystr" {
-		t.Errorf("GETRANGE: expected [mystr], got %v", keys)
-	}
-}
-
-func TestExtractRedisKeys_MultiKey(t *testing.T) {
-	cmd := makeCmd("MGET", "a", "b", "c")
-	keys := extractRedisKeys(cmd)
-	if len(keys) != 3 {
-		t.Fatalf("expected 3 keys, got %d: %v", len(keys), keys)
-	}
-	want := []string{"a", "b", "c"}
-	for i, k := range keys {
-		if k != want[i] {
-			t.Errorf("key[%d] = %q, want %q", i, k, want[i])
-		}
-	}
-}
-
-func TestExtractRedisKeys_MultiKeyExists(t *testing.T) {
-	cmd := makeCmd("EXISTS", "k1", "k2", "k3")
-	keys := extractRedisKeys(cmd)
-	if len(keys) != 3 {
-		t.Fatalf("EXISTS: expected 3 keys, got %d: %v", len(keys), keys)
-	}
-}
-
-func TestExtractRedisKeys_NumKeysPattern(t *testing.T) {
-	// ZDIFF numkeys key [key ...]
-	cmd := makeCmd("ZDIFF", 2, "zs1", "zs2")
-	cmd.(*Cmd).SetFirstKeyPos(2)
-	keys := extractRedisKeys(cmd)
-	if len(keys) != 2 || keys[0] != "zs1" || keys[1] != "zs2" {
-		t.Errorf("ZDIFF: expected [zs1 zs2], got %v", keys)
-	}
-
-	// SINTERCARD numkeys key [key ...] LIMIT limit
-	cmd = makeCmd("SINTERCARD", 2, "s1", "s2", "LIMIT", 10)
-	keys = extractRedisKeys(cmd)
-	if len(keys) != 2 || keys[0] != "s1" || keys[1] != "s2" {
-		t.Errorf("SINTERCARD: expected [s1 s2], got %v", keys)
-	}
-}
-
-func TestExtractRedisKeys_LCS(t *testing.T) {
-	cmd := makeCmd("LCS", "key1", "key2")
-	keys := extractRedisKeys(cmd)
-	if len(keys) != 2 || keys[0] != "key1" || keys[1] != "key2" {
-		t.Errorf("LCS: expected [key1 key2], got %v", keys)
-	}
-}
-
-func TestExtractRedisKeys_JSONMGet(t *testing.T) {
-	// JSON.MGET key [key ...] path
-	cmd := makeCmd("JSON.MGET", "j1", "j2", "$.name")
-	keys := extractRedisKeys(cmd)
-	if len(keys) != 2 || keys[0] != "j1" || keys[1] != "j2" {
-		t.Errorf("JSON.MGET: expected [j1 j2], got %v", keys)
-	}
-}
-
-func TestExtractRedisKeys_KeylessCommand(t *testing.T) {
-	cmd := makeCmd("ping")
-	keys := extractRedisKeys(cmd)
-	if keys != nil {
-		t.Errorf("expected nil for keyless command, got %v", keys)
-	}
-}
-
-func TestIsCacheable_SortRO_ByGetExcluded(t *testing.T) {
-	// Plain SORT_RO reads only the sorted key: cacheable.
-	if cmd := makeCmd("sort_ro", "mylist", "LIMIT", "0", "10", "ALPHA"); !isCacheable(cmd) {
-		t.Error("plain SORT_RO should be cacheable")
-	}
-	// BY/GET forms read pattern-derived keys the reverse index cannot cover:
-	// their invalidations would be dropped, serving stale results forever.
-	if cmd := makeCmd("sort_ro", "mylist", "BY", "weight_*"); isCacheable(cmd) {
-		t.Error("SORT_RO ... BY must not be cacheable")
-	}
-	if cmd := makeCmd("sort_ro", "mylist", "get", "obj_*"); isCacheable(cmd) {
-		t.Error("SORT_RO ... GET must not be cacheable (case-insensitive)")
-	}
-	if cmd := makeCmd("sort_ro", "mylist", "LIMIT", "0", "10", "By", "weight_*", "ALPHA"); isCacheable(cmd) {
-		t.Error("SORT_RO with BY among other options must not be cacheable")
-	}
-	by := "BY"
-	if cmd := makeCmd("sort_ro", "mylist", &by, "weight_*"); isCacheable(cmd) {
-		t.Error("SORT_RO with pointer-encoded BY must not be cacheable")
-	}
-}
