+++ internal/pool/pool_test.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"net"
+	"reflect"
 	"sync"
 	"sync/atomic"
 	"testing"
@@ -324,6 +325,214 @@ func TestMaxActiveConnsStillReturnsPoolExhausted(t *testing.T) {
 	if _, err := connPool.Get(ctx); !errors.Is(err, pool.ErrPoolExhausted) {
 		t.Fatalf("second Get: got %v, want ErrPoolExhausted", err)
 	}
+}
+
+func TestDrainIdleConnsReportsBorrowedConnectionStates(t *testing.T) {
+	connPool := pool.NewConnPool(&pool.Options{
+		Dialer:             dummyDialer,
+		PoolSize:           1,
+		MaxConcurrentDials: 1,
+		PoolTimeout:        time.Second,
+		ConnMaxIdleTime:    -1,
+	})
+	t.Cleanup(func() { _ = connPool.Close() })
+
+	ctx := context.Background()
+	cn, err := connPool.Get(ctx)
+	if err != nil {
+		t.Fatalf("Get: %v", err)
+	}
+	cn.GetStateMachine().Transition(pool.StateInUse)
+	connPool.Put(ctx, cn)
+
+	type event struct {
+		delta int
+		state string
+	}
+	var mu sync.Mutex
+	var events []event
+	pool.SetAllMetricCallbacks(&pool.MetricCallbacks{
+		ConnectionCount: func(_ context.Context, delta int, _ *pool.Conn, state string, _ bool) {
+			mu.Lock()
+			events = append(events, event{delta: delta, state: state})
+			mu.Unlock()
+		},
+	})
+	defer pool.SetAllMetricCallbacks(nil)
+
+	var state pool.DrainState
+	connPool.DrainIdleConns(ctx, &state, func(*pool.Conn) error { return nil })
+
+	mu.Lock()
+	defer mu.Unlock()
+	want := []event{
+		{delta: -1, state: pool.MetricStateIdle},
+		{delta: 1, state: pool.MetricStateUsed},
+		{delta: -1, state: pool.MetricStateUsed},
+		{delta: 1, state: pool.MetricStateIdle},
+	}
+	if !reflect.DeepEqual(events, want) {
+		t.Fatalf("connection state events: got %#v, want %#v", events, want)
+	}
+}
+
+func TestDrainIdleConnsSkipsUninitializedMinIdleConns(t *testing.T) {
+	connPool := pool.NewConnPool(&pool.Options{
+		Dialer:             dummyDialer,
+		PoolSize:           8,
+		MinIdleConns:       8,
+		MaxConcurrentDials: 8,
+		PoolTimeout:        time.Second,
+		ConnMaxIdleTime:    -1,
+	})
+	t.Cleanup(func() { _ = connPool.Close() })
+
+	deadline := time.Now().Add(time.Second)
+	for connPool.Len() != 8 && time.Now().Before(deadline) {
+		time.Sleep(time.Millisecond)
+	}
+	if got := connPool.Len(); got != 8 {
+		t.Fatalf("MinIdleConns: got %d, want 8", got)
+	}
+
+	var state pool.DrainState
+	called := 0
+	connPool.DrainIdleConns(context.Background(), &state, func(*pool.Conn) error {
+		called++
+		return nil
+	})
+	if called != 0 {
+		t.Fatalf("drained %d uninitialized MinIdleConns, want 0", called)
+	}
+}
+
+func TestDrainIdleConnsVisitsSnapshotOnce(t *testing.T) {
+	for _, fifo := range []bool{false, true} {
+		t.Run(fmt.Sprintf("fifo=%v", fifo), func(t *testing.T) {
+			const size = 16
+			connPool := pool.NewConnPool(&pool.Options{
+				Dialer:             dummyDialer,
+				PoolFIFO:           fifo,
+				PoolSize:           size,
+				MaxConcurrentDials: size,
+				PoolTimeout:        time.Second,
+				ConnMaxIdleTime:    -1,
+			})
+			t.Cleanup(func() { _ = connPool.Close() })
+
+			ctx := context.Background()
+			conns := make([]*pool.Conn, size)
+			for i := range conns {
+				cn, err := connPool.Get(ctx)
+				if err != nil {
+					t.Fatalf("Get %d: %v", i, err)
+				}
+				cn.GetStateMachine().Transition(pool.StateInUse)
+				conns[i] = cn
+			}
+			for _, cn := range conns {
+				connPool.Put(ctx, cn)
+			}
+
+			seen := make(map[uint64]int, size)
+			drainCtx, cancel := context.WithCancel(ctx)
+			var state pool.DrainState
+			connPool.DrainIdleConns(drainCtx, &state, func(cn *pool.Conn) error {
+				seen[cn.GetID()]++
+				cancel()
+				return nil
+			})
+			connPool.DrainIdleConns(ctx, &state, func(cn *pool.Conn) error {
+				seen[cn.GetID()]++
+				return nil
+			})
+
+			if len(seen) != size {
+				t.Fatalf("drained %d unique connections, want %d", len(seen), size)
+			}
+			for id, count := range seen {
+				if count != 1 {
+					t.Fatalf("connection %d drained %d times, want 1", id, count)
+				}
+			}
+		})
+	}
+}
+
+type acceptingPoolHook struct{}
+
+func (acceptingPoolHook) OnGet(context.Context, *pool.Conn, bool) (bool, error) {
+	return true, nil
+}
+
+func (acceptingPoolHook) OnPut(context.Context, *pool.Conn) (bool, bool, error) {
+	return true, false, nil
+}
+
+func (acceptingPoolHook) OnRemove(context.Context, *pool.Conn, error) {}
+
+func TestCreateTimeWaitsForHookInitialization(t *testing.T) {
+	var createCalls atomic.Int32
+	pool.SetAllMetricCallbacks(&pool.MetricCallbacks{
+		ConnectionCreateTime: func(context.Context, time.Duration, *pool.Conn) {
+			createCalls.Add(1)
+		},
+	})
+	defer pool.SetAllMetricCallbacks(nil)
+
+	connPool := pool.NewConnPool(&pool.Options{
+		Dialer:             dummyDialer,
+		PoolSize:           1,
+		MaxConcurrentDials: 1,
+		PoolTimeout:        time.Second,
+	})
+	t.Cleanup(func() { _ = connPool.Close() })
+	connPool.AddPoolHook(acceptingPoolHook{})
+
+	cn, err := connPool.Get(context.Background())
+	if err != nil {
+		t.Fatalf("Get: %v", err)
+	}
+	if cn.IsInited() {
+		t.Fatal("test hook unexpectedly initialized the connection")
+	}
+	if got := createCalls.Load(); got != 0 {
+		t.Fatalf("creation metric calls before initialization: got %d, want 0", got)
+	}
+	cn.GetStateMachine().Transition(pool.StateInUse)
+	connPool.Put(context.Background(), cn)
+}
+
+func TestConnectionWaitTimeExcludesDial(t *testing.T) {
+	const dialDelay = 100 * time.Millisecond
+	var waitDuration atomic.Int64
+	pool.SetAllMetricCallbacks(&pool.MetricCallbacks{
+		ConnectionWaitTime: func(_ context.Context, duration time.Duration, _ *pool.Conn) {
+			waitDuration.Store(int64(duration))
+		},
+	})
+	defer pool.SetAllMetricCallbacks(nil)
+
+	connPool := pool.NewConnPool(&pool.Options{
+		Dialer: func(context.Context) (net.Conn, error) {
+			time.Sleep(dialDelay)
+			return dummyDialer(context.Background())
+		},
+		PoolSize:           1,
+		MaxConcurrentDials: 1,
+		PoolTimeout:        time.Second,
+	})
+	t.Cleanup(func() { _ = connPool.Close() })
+
+	cn, err := connPool.Get(context.Background())
+	if err != nil {
+		t.Fatalf("Get: %v", err)
+	}
+	if got := time.Duration(waitDuration.Load()); got >= dialDelay/2 {
+		t.Fatalf("connection wait metric included dial time: got %v, dial took %v", got, dialDelay)
+	}
+	cn.GetStateMachine().Transition(pool.StateInUse)
+	connPool.Put(context.Background(), cn)
 }
 
 var _ = Describe("MinIdleConns", func() {
